Moving to v1 module support
