+++ go.mod
@@ -1,0 +1,12 @@
+module github.com/Datera/go-sdk
+
+require (
+	github.com/Datera/go-udc v1.1.0
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/google/uuid v1.1.0
+	github.com/levigross/grequests v0.0.0-20181123014746-f3f67e7783bb
+	github.com/mitchellh/go-homedir v1.0.0 // indirect
+	github.com/mitchellh/mapstructure v1.1.2
+	github.com/sirupsen/logrus v1.3.0
+	golang.org/x/net v0.0.0-20190108155000-395948e2f546 // indirect
+)
